# Asynchronous detectors and Souffle execution