--- a/src/detectors/builtin/preferAugmentedAssign.ts
+++ b/src/detectors/builtin/preferAugmentedAssign.ts
@@ -53,7 +53,7 @@ const DontSuggestKinds = new Set<string>([
  * ```
  */
 export class PreferAugmentedAssign extends Detector {
-  check(cu: CompilationUnit): MistiTactWarning[] {
+  async check(cu: CompilationUnit): Promise<MistiTactWarning[]> {
     return cu.ast.getProgramEntries().reduce((acc, node) => {
       return acc.concat(
         foldStatements(node, [] as MistiTactWarning[], (acc, expr) => {