--- a/src/detectors/builtin/unboundLoops.ts
+++ b/src/detectors/builtin/unboundLoops.ts
@@ -50,13 +50,13 @@ import {
  * ```
  */
 export class UnboundLoops extends SouffleDetector {
-  check(cu: CompilationUnit): MistiTactWarning[] {
-    const program = new Context<SrcInfo>(this.id);
+  async check(cu: CompilationUnit): Promise<MistiTactWarning[]> {
+    const program = this.createSouffleContext(cu);
     this.addDecls(program);
     this.addRules(program);
     this.addConstantConstraints(cu, program);
     this.addConstraints(cu, program);
-    return this.executeSouffle(program, (fact) => {
+    return await this.executeSouffle(program, (fact) => {
       if (fact.data === undefined) {
         throw new Error(`AST position for fact ${fact} is not available`);
       }