--- a/src/detectors/builtin/zeroAddress.ts
+++ b/src/detectors/builtin/zeroAddress.ts
@@ -42,7 +42,7 @@ import { AstExpression } from "@tact-lang/compiler/dist/grammar/ast";
  * ```
  */
 export class ZeroAddress extends Detector {
-  check(cu: CompilationUnit): MistiTactWarning[] {
+  async check(cu: CompilationUnit): Promise<MistiTactWarning[]> {
     return cu.ast.getProgramEntries().reduce((acc, node) => {
       return acc.concat(
         foldExpressions(node, [] as MistiTactWarning[], (acc, expr) => {