--- a/src/internals/souffle/executor/index.ts
+++ b/src/internals/souffle/executor/index.ts
@@ -1,0 +1,7 @@
+export {
+  SyncExecutor,
+  AsyncExecutor,
+  SouffleExecutionResult,
+  SouffleExecutorParams,
+} from "./executor";
+export { ParsedSouffleOutput } from "./results";