--- a/src/internals/souffle/executor/index.ts
+++ b/src/internals/souffle/executor/index.ts
@@ -1,4 +1,5 @@
 export {
+  Executor,
   SyncExecutor,
   AsyncExecutor,
   SouffleExecutionResult,