Closes #27

- [x] I have updated `CHANGELOG.md`
- [ ] I have added tests to demonstrate the contribution is correctly implemented
- [x] No test failures were reported when running `yarn test-all`
- [x] I did not do unrelated and/or undiscussed refactorings