@jubnzv 2024-08-31
It gives a small performance boost comparing with `master`.

`master` benchmark results on `sample-jetton.tact`:
| Detector Name | Result |
|---------------|--------|
| **All Detectors** | 0.49 ops/sec (~2033.16 ms/run) ±9.56% |
| `DivideBeforeMultiply` | 0.54 ops/sec (~1837.80 ms/run) ±0.98% |
| `ReadOnlyVariables` | 0.54 ops/sec (~1849.14 ms/run) ±4.56% |
| `NeverAccessedVariables` | 0.56 ops/sec (~1777.49 ms/run) ±1.55% |
| `UnboundLoops` | 0.55 ops/sec (~1822.80 ms/run) ±3.80% |
| `ZeroAddress` | 0.57 ops/sec (~1755.86 ms/run) ±0.74% |
| `ConstantAddress` | 0.56 ops/sec (~1782.98 ms/run) ±3.75% |
| `BranchDuplicate` | 0.57 ops/sec (~1757.69 ms/run) ±1.18% |
| `DumpIsUsed` | 0.56 ops/sec (~1792.60 ms/run) ±4.53% |
| `FieldDoubleInit` | 0.57 ops/sec (~1759.16 ms/run) ±1.31% |
| `PreferAugmentedAssign` | 0.57 ops/sec (~1763.16 ms/run) ±4.55% |

This PR benchmark results on `sample-jetton.tact`:
| Detector Name | Result |
|---------------|--------|
| **All Detectors** | 0.54 ops/sec (~1865.07 ms/run) ±6.96% |
| `DivideBeforeMultiply` | 0.56 ops/sec (~1789.02 ms/run) ±4.39% |
| `ReadOnlyVariables` | 0.58 ops/sec (~1727.20 ms/run) ±0.66% |
| `NeverAccessedVariables` | 0.58 ops/sec (~1730.55 ms/run) ±3.65% |
| `UnboundLoops` | 0.58 ops/sec (~1715.04 ms/run) ±0.85% |
| `ZeroAddress` | 0.59 ops/sec (~1693.81 ms/run) ±0.71% |
| `ConstantAddress` | 0.56 ops/sec (~1782.88 ms/run) ±6.80% |
| `BranchDuplicate` | 0.53 ops/sec (~1901.73 ms/run) ±5.95% |
| `DumpIsUsed` | 0.56 ops/sec (~1795.69 ms/run) ±1.22% |
| `FieldDoubleInit` | 0.55 ops/sec (~1813.77 ms/run) ±3.05% |
| `PreferAugmentedAssign` | 0.57 ops/sec (~1747.10 ms/run) ±1.01% |

But the performance difference should be more remarkable when using more complicated Souffle detectors or introducing any parallelizable logic to other detectors (e.g. filesystem operations, API calls, programs execution).